+++ packages/app/features/debug/lib/src/ui/debug_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
+import 'package:packages_app_features_debug/src/ui/components/route_dropdown_menu.dart';
 
 class DebugPage extends StatelessWidget {
   const DebugPage({
@@ -8,69 +8,20 @@ class DebugPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final router = GoRouter.of(context);
-    final goRoutes = router.configuration.routes.toGoRoutes();
-    return Scaffold(
+    return const Scaffold(
       body: CustomScrollView(
         slivers: [
-          const SliverAppBar.large(
+          SliverAppBar.large(
             title: Text('Debug Page'),
           ),
           SliverToBoxAdapter(
             child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16),
-              child: LayoutBuilder(
-                builder: (context, constraints) {
-                  return DropdownMenu<GoRoute>(
-                    width: constraints.maxWidth,
-                    dropdownMenuEntries: goRoutes
-                        .map(
-                          (goRoute) {
-                            // debug routes are excluded
-                            if (goRoute.path.contains('debug')) {
-                              return null;
-                            }
-                            return DropdownMenuEntry<GoRoute>(
-                              value: goRoute,
-                              label: goRoute.path,
-                            );
-                          },
-                        )
-                        .nonNulls
-                        .toList(),
-                    onSelected: (goRoute) {
-                      if (goRoute == null) {
-                        return;
-                      }
-                      router.go(goRoute.path);
-                    },
-                  );
-                },
-              ),
+              padding: EdgeInsets.symmetric(horizontal: 16),
+              child: RouteDropdownMenu(),
             ),
           ),
         ],
       ),
     );
-  }
-}
-
-extension _ToGoRoutes on List<RouteBase> {
-  List<GoRoute> toGoRoutes() {
-    final goRoutes = <GoRoute>[];
-    for (final route in this) {
-      switch (route) {
-        case GoRoute():
-          final childRoutes = route.routes;
-          if (childRoutes.isEmpty) {
-            goRoutes.add(route);
-          } else {
-            goRoutes.addAll(childRoutes.toGoRoutes());
-          }
-        case ShellRoute() || StatefulShellRoute():
-          goRoutes.addAll(route.routes.toGoRoutes());
-      }
-    }
-    return goRoutes;
   }
 }
